Autocorrect data after checking session # and session type
checkSessionDates.m:  if incorrect session numbers are found, medPC files are updated and a new master table is created and saved. if incorrect session types are found, the master table will be updated and saved.

also added an experiment column to the master table, here. might make more sense to move it to createNewMasterTable.m

corrected files are saved.