Indexing system, reinstatement latency, improved flexibility, hdf5 saving, CC license
Removed files from run2 that we're not interested in analyzing (progressive ratio and extinction sessions that occurred after behavioral economics sessions)

createMasterTable adds Run & Experiment columns to mT

runNum is now set with a string. multiple runs can be specified with underscores interleaved like '1_2_4' if, say, you wanted SA data from all runs except run 3.

added an "all" option to runType. In this case, all analyses are run on subsets of the data that they apply to.

implemented a new indexing system for subsetting data in mT to be analyzed. rather than cropping the mT table to the data to analyze, an index structure is created. The index includes categories for each runType with the runs to be included, as well as an "all" category for areas of the code where we want all data processed regardless of type.

Updated definition of allLatency and Latency for Reinstatement sessions. Now instead of having no latency data on reinstatement, latency is calculated between all active lever presses and the subsequent head entries.

un-hardcoded within-session behavioral analysis drug intake. this now reads from Experiment Key. note that within-session BE intake results may not exactly match the Behavioral-Economics analysis section results, because the latter reads from the "Intake Canonical" sheet.

cleaned up how tables were being saved, and made them save according to the data used for a given analysis  (i.e. crops mT to the data subset used). added table saving to within-session behavior analysis.

cleaned up Behavioral Economics analysis section & added firstHour data

added runType flexibility to individual variability/susceptibility modeling section. Still needs to be cleaned up quite a bit, but it won't break on BE and SA data now.

added scripts for saving tables as hdf5s and loading them back as matlab tables with the exact same classes and data shapes. this still needs to be implemented in the createMasterTable and loading section of the main analysis script, but it can be generated with the test_loadHDF5 script. had to remove the output .hdf5 itself from the repo because it was 300MB for some reason...

added a non-commercial-no-derivatives creative commons license.